+++ CHANGELOG.md
@@ -46,6 +46,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - When auto-checking for updates, the program would consider there are updates if your version number is greater than the one from GitHub - it's
   unlikely that the end user ever saw this, but it's now fixed.
 
+## Unreleased
+
+### Added
+
+- Added `migrate` command to move a project from one package to another.
+
 ## [1.1.0] 19-10-2024
 
 ### Added